backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
xgboost==2.1.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
pyjwt==2.9.0
pandas==2.2.2
scikit-learn==1.5.2
SQLAlchemy==2.0.36
asyncpg==0.29.0
